+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 ### Added
 - `duplicate` methods to `Array` and `Hash`
 - More `PartialEq` and `PartialOrd` implementations to `Float` and `Integer`
+- `meta` module with Ruby's metadata
 
 ### Changed
 - Where `Float` and `Integer` are located; both now reside in a `num` module